src/djangorestapi/requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.4
black==21.5b1
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.8
Django==3.2
django-cors-headers==3.7.0
django-crontab==0.7.1
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.24
mysqlclient==2.0.3
openapi-codec==1.3.2
packaging==20.9
pathspec==0.8.1
Pillow==8.2.0
protobuf==3.15.8
pyparsing==2.4.7
python-telegram-bot==13.4.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
uWSGI==2.0.19.1
